django-cassandra-test/requirements/local.txt
cassandra-driver==3.12.0
Django==1.11.7
django-cassandra-engine==1.3.0
django-environ==0.4.4
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
numpy==1.13.3
pandas==0.21.0
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.3
scipy==1.0.0
six==1.11.0
uuid==1.30
